+++ lib/td_dq/rules.ex
@@ -285,7 +285,7 @@ defmodule TdDq.Rules do
 
   def get_cached_content(%{} = content, type) when is_binary(type) do
     case TemplateCache.get_by_name!(type) do
-      template = %{} -> Format.enrich_content_values(content, template)
+      template = %{} -> Format.enrich_content_values(content, template, [:system, :hierarchy])
       _ -> content
     end
   end